code/interface/requirements.txt
dash==2.7.0
dash-bootstrap-components==1.0.3
dash-leaflet==0.1.18
numpy==1.22.4
pandas==1.4.2
plotly==5.6.0
pyserial==3.5

matplotlib
